feat: add arXiv source processor
